% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_lipidomics_day14}
\alias{rat_lipidomics_day14}
\title{Rat Plasma Lipidomics (Day 14)}
\format{
\subsection{\code{rat_lipidomics_day0}}{

A data frame with 60 rows and 302 columns where rows correspond to observations
and columns correspond to features
}
}
\source{
University of Michigan Metabolomics Core
}
\usage{
rat_lipidomics_day14
}
\description{
A dataframe of raw peak areas from untargeted lipidomics analysis after correcting
for batch, drift, and missingness. Features with high variability or unknown lipid
identification have been removed.
}
\keyword{datasets}
