% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_lipidomics_group_key}
\alias{rat_lipidomics_group_key}
\title{Rat Lipidomics Group Key Identifier}
\format{
\subsection{\code{rat_lipidomics_group_key}}{

A data frame with 302 rows and 2 columns
\describe{
\item{Lipid}{Feature name, which corresponds to column names for data files}
\item{Cluster}{Group identifier for PCA feature reduction analysis }
}
}
}
\source{
University of Michigan Metabolomics Core
}
\usage{
rat_lipidomics_group_key
}
\description{
A dataframe containing the feature names and their corresponding groups IDs
}
\keyword{datasets}
