% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_meta_data}
\alias{rat_meta_data}
\title{Rat Phenotype Data}
\format{
\subsection{\code{rat_meta_data}}{

A data frame with 60 rows and 4 columns
\describe{
\item{Name}{Rat ID for linking data to feature data}
\item{Line}{High Capacity Runner (HCR) or Low Capacity Runner (LCR) rat line identifier}
\item{Sex}{Female or male rat sex identifer}
\item{Diet}{Control, Carbohydrate, or Fat rat diet identifier}
}
}
}
\source{
University of Michigan Metabolomics Core
}
\usage{
rat_meta_data
}
\description{
A dataframe containing the rat meta and phenotype data use for linear regression
modeling
}
\keyword{datasets}
