% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reduce_features_by_pca}
\alias{reduce_features_by_pca}
\title{Feature Reduction by PCA}
\usage{
reduce_features_by_pca(
  feature_data = NULL,
  feature_group_key = NULL,
  norm = "none"
)
}
\arguments{
\item{feature_data}{Dataframe with columns as features and rows as observations.
This function does not perform any data cleaning or normalization. For metabolomics
data, it is expected that data is log2 normalized. If \code{NA} features are present, these
will be excluded from PCA analysis and the \code{NA} value for the will be retained in
the output object.}

\item{feature_group_key}{Dataframe with column 1 as features and column 2 as group ID.
Column names in \code{feature_data} must have exact matches in column 1 of \code{feature_group_key} to
be included in the grouping analysis.}

\item{norm}{One of \code{"none"} (default) or \code{"scale"} to indicate weather features should be scaled
prior to weighing. This is helpful for heatmaps. Note that PCA will be performed on the
provided \code{feature_data} values.}
}
\value{
A matrix with columns as groups (provided group IDs) and rows as observations.
}
\description{
This function accepts feature data and a predefined feature group key to
compute a weighed group value. Within each predefined group, features are
weighed by PCA, multiplied by the absolute value of the PCA loading factor to
determine its relative contribution to variance, and summed to generate a group
value. Each observation is assigned a group value in the dataframe output.
}
